requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
python-dotenv==1.0.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1
webauthn==1.12.0
cbor2==5.4.6
pycryptodome==3.19.0
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
Alembic==1.12.1
Faker==20.1.0
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
redis==5.0.1